+++ src/main/java/controller/RestController.java
@@ -6,7 +6,7 @@ import model.HttpResponse;
 import model.enums.HttpStatusCode;
 import model.enums.Method;
 import service.FileService;
-import service.ResponseService;
+import mapper.ResponseMapper;
 import service.UserService;
 
 import java.io.FileNotFoundException;
@@ -17,16 +17,16 @@ import static util.StringUtils.NO_CONTENT;
 public class RestController {
     private final FileService fileService;
     private final UserService userService;
-    private final ResponseService responseService;
+    private final ResponseMapper responseMapper;
 
     public RestController() {
         fileService = new FileService();
         userService = new UserService();
-        responseService = new ResponseService();
+        responseMapper = new ResponseMapper();
     }
 
     public HttpResponse route(HttpRequest httpRequest) throws FileNotFoundException {
-        HttpResponse response = responseService.createNotFoundResponse(httpRequest);
+        HttpResponse response = responseMapper.createNotFoundResponse(httpRequest);
         if (httpRequest.match(Method.GET) && httpRequest.endsWithHtml()) {
             response = sendNotRestfulResponse(httpRequest);
         }
@@ -41,10 +41,10 @@ public class RestController {
         try {
             UserFormRequestDto userFormRequestDto = request.paramsToDto();
             userService.createByForm(userFormRequestDto);
-            return responseService
+            return responseMapper
                     .createHttpResponse(request, HttpStatusCode.CREATED, NO_CONTENT);
         } catch (Exception e) {
-            return responseService.createBadRequestResponse(request);
+            return responseMapper.createBadRequestResponse(request);
         }
     }
 
@@ -55,10 +55,10 @@ public class RestController {
     private HttpResponse getHttpResponse(HttpRequest request, String path) {
         try {
             String fileContents = fileService.openFile(path);
-            return responseService
+            return responseMapper
                     .createHttpResponse(request, HttpStatusCode.OK, fileContents);
         } catch (FileNotFoundException e) {
-            return responseService.createBadRequestResponse(request);
+            return responseMapper.createBadRequestResponse(request);
         }
     }
 }
